Added method to check player presence in game session
The code now includes the isPlayerInSession method within the GameSession aggregate root, which checks whether a specific player is in the session or not. Also, the CurrentGameSessionMysqlGetter class has been refactored to use this new method, instead of directly accessing the teams attribute. This improves encapsulation and reusability of the code for checking whether a player is in a session or not.